Classes Wall and Ball were separated.
